Move CallWindow class to its own file
